+++ src/features/welcome/welcome.tsx
@@ -1,17 +1,18 @@
 import { Box, Typography } from "@mui/material"
+import styles from '../../styles.js'
 
-const styles = {
-  color: {
-    background: 'radial-gradient(circle at center, rgba(65, 162, 72, 0.4), rgba(65, 162, 72, 0.1))',
-    borderRadius: '15px',
-    paddingLeft: '30px',
-    paddingRight: '30px',
-    paddingTop: '60px',
-    paddingBottom: '60px',
-    display: 'flex',
-    justifyItems: 'center',
-  },
-};
+// const styles = {
+//   color: {
+//     background: 'radial-gradient(circle at center, rgba(65, 162, 72, 0.4), rgba(65, 162, 72, 0.1))',
+//     borderRadius: '15px',
+//     paddingLeft: '30px',
+//     paddingRight: '30px',
+//     paddingTop: '60px',
+//     paddingBottom: '60px',
+//     display: 'flex',
+//     justifyItems: 'center',
+//   },
+// };
 
 function WelcomeFeature() {
   return (
