+++ src/styles.js
@@ -1,0 +1,14 @@
+const styles = {
+  color: {
+    background: 'radial-gradient(circle at center, rgba(65, 162, 72, 0.4), rgba(65, 162, 72, 0.1))',
+    borderRadius: '15px',
+    paddingLeft: '30px',
+    paddingRight: '30px',
+    paddingTop: '60px',
+    paddingBottom: '60px',
+    display: 'flex',
+    justifyItems: 'center',
+  },
+};
+
+export default styles;
